# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/lessons/{lessonId}/sections": {
      "get": {
        "description": "레슨 내 섹션 목록과 각 섹션의 진행 상태를 반환합니다.",
        "operationId": "LearningController_lessonSections",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "description": "레슨 ID",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "최상위 레슨 제목 필드명은 `title`(Swagger 예전 `lessonTitle` 아님). `siblingLessons`·섹션별 `totalPages`·`currentPage`·`progressPercent`·`hasContent`·`orderNum` 포함.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "isSuccess": true,
                    "code": "200",
                    "message": "요청이 성공했습니다.",
                    "data": {
                      "courseId": 1,
                      "lessonId": 2,
                      "title": "기초 문법 1",
                      "subtitle": null,
                      "siblingLessons": [
                        {
                          "lessonId": 1,
                          "title": "인사하기",
                          "orderNum": 1
                        },
                        {
                          "lessonId": 2,
                          "title": "기초 문법 1",
                          "orderNum": 2
                        }
                      ],
                      "overallProgressPercent": 50,
                      "sections": [
                        {
                          "sectionId": 5,
                          "type": "GRAMMAR",
                          "title": "동사 활용",
                          "totalPages": 5,
                          "orderNum": 1,
                          "currentPage": 5,
                          "progressPercent": 100,
                          "isCompleted": true,
                          "hasContent": true
                        },
                        {
                          "sectionId": 6,
                          "type": "VOCAB",
                          "title": "단어 카드",
                          "totalPages": 3,
                          "orderNum": 2,
                          "currentPage": 1,
                          "progressPercent": 33,
                          "isCompleted": false,
                          "hasContent": true
                        }
                      ]
                    },
                    "timestamp": "2026-04-05T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "레슨 없음",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "isSuccess": false,
                    "code": "404",
                    "message": "레슨을 찾을 수 없습니다.",
                    "data": null,
                    "errorCode": "LESSON_NOT_FOUND",
                    "timestamp": "2026-04-05T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "레슨 섹션 목록 조회",
        "tags": [
          "학습 (Learning)"
        ]
      }
    },
...
    "/admin/overview": {
      "get": {
        "operationId": "AdminController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "전체 콘텐츠 트리 조회 (코스 → 레슨 → 섹션 + 콘텐츠 개수)",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/contents": {
      "get": {
        "operationId": "AdminController_getSectionContents",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "섹션 콘텐츠 상세 조회 (카드·자료·문제, 정답 포함)",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/courses": {
      "post": {
        "operationId": "AdminController_createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "코스 생성",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/courses/{courseId}": {
      "patch": {
        "operationId": "AdminController_patchCourse",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "코스 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteCourse",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "코스 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/courses/{courseId}/lessons": {
      "post": {
        "operationId": "AdminController_createLesson",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "레슨 생성",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/lessons/{lessonId}": {
      "patch": {
        "operationId": "AdminController_patchLesson",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "레슨 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteLesson",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "레슨 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/lessons/{lessonId}/sections": {
      "post": {
        "operationId": "AdminController_createSection",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "섹션 생성",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}": {
      "patch": {
        "operationId": "AdminController_patchSection",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "섹션 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteSection",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "섹션 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/cards": {
      "post": {
        "operationId": "AdminController_createCard",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "단어 카드 단건 추가",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/cards/bulk": {
      "post": {
        "operationId": "AdminController_createCardsBulk",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCardsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "단어 카드 일괄 추가",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/cards/{cardId}": {
      "patch": {
        "operationId": "AdminController_patchCard",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "카드 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteCard",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "카드 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/materials": {
      "post": {
        "operationId": "AdminController_createMaterial",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "학습 자료 추가",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/materials/{materialId}": {
      "patch": {
        "operationId": "AdminController_patchMaterial",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMaterialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "학습 자료 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteMaterial",
        "parameters": [
          {
            "name": "materialId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "학습 자료 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/questions": {
      "post": {
        "operationId": "AdminController_createQuestion",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "문제 추가",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/questions/{questionId}": {
      "patch": {
        "operationId": "AdminController_patchQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "문제 수정",
        "tags": [
          "관리자 (Admin)"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "문제 삭제",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/badges/refresh": {
      "post": {
        "operationId": "AdminController_refreshBadgeCache",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "배지 캐시 갱신 (Redis Hash 재동기화)",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
    "/admin/sections/{sectionId}/audio-upload-url": {
      "post": {
        "operationId": "AdminController_presignedAudio",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAudioPresignedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "오디오 S3 presigned URL 발급",
        "tags": [
          "관리자 (Admin)"
        ]
      }
    },
...
  },
  "info": {
    "title": "DOJEON API",
    "description": "외국인 대상 한국어 학습 플랫폼 API\n\n**응답 형식**\n- 성공: `{ isSuccess, code, message, data, timestamp }`\n- 실패: `{ isSuccess, code, message, data, errorCode?, timestamp }`\n\n**인증**\n- 보호된 엔드포인트는 `Authorization: Bearer <accessToken>`\n- Swagger에서는 우측 **Authorize**에 토큰만 넣으면 됩니다 (Bearer 접두어 없이).\n\n**프론트엔드용 스펙 파일**\n- OpenAPI JSON: `/docs-json` (Postman·Orval 등)\n- OpenAPI YAML: `/docs-yaml`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "문서를 연 호스트와 동일 (로컬·배포 공통)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "orderNum"
        ]
      },
      "PatchCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "orderNum"
        ]
      },
      "PatchLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "orderNum": {
            "type": "number"
          }
        }
      },
      "CreateSectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "VOCAB",
              "GRAMMAR",
              "READING",
              "LISTENING",
              "QUIZ"
            ]
          },
          "title": {
            "type": "string"
          },
          "totalPages": {
            "type": "number"
          },
          "orderNum": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "title",
          "totalPages",
          "orderNum"
        ]
      },
      "PatchSectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "VOCAB",
              "GRAMMAR",
              "READING",
              "LISTENING",
              "QUIZ"
            ]
          },
          "title": {
            "type": "string"
          },
          "totalPages": {
            "type": "number"
          },
          "orderNum": {
            "type": "number"
          }
        }
      },
      "CreateCardDto": {
        "type": "object",
        "properties": {
          "wordFront": {
            "type": "string"
          },
          "wordBack": {
            "type": "string"
          },
          "notes": {
            "type": "object",
            "description": "영문 노트(카드 뒷면 보조 설명)"
          },
          "locales": {
            "type": "object",
            "description": "로케일별 번역·노트. 예: `{ \"he\": { \"back\": \"…\", \"notes\": \"…\" } }`",
            "additionalProperties": true
          },
          "audioUrl": {
            "type": "string"
          },
          "sequence": {
            "type": "number"
          }
        },
        "required": [
          "wordFront",
          "wordBack",
          "sequence"
        ]
      },
      "BulkCardsItemDto": {
        "type": "object",
        "properties": {
          "wordFront": {
            "type": "string"
          },
          "wordBack": {
            "type": "string"
          },
          "notes": {
            "type": "object"
          },
          "locales": {
            "type": "object",
            "additionalProperties": true
          },
          "audioUrl": {
            "type": "string"
          },
          "sequence": {
            "type": "number"
          }
        },
        "required": [
          "wordFront",
          "wordBack",
          "sequence"
        ]
      },
      "BulkCardsDto": {
        "type": "object",
        "properties": {
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkCardsItemDto"
            }
          }
        },
        "required": [
          "cards"
        ]
      },
      "PatchCardDto": {
        "type": "object",
        "properties": {
          "wordFront": {
            "type": "string"
          },
          "wordBack": {
            "type": "string"
          },
          "notes": {
            "type": "object"
          },
          "locales": {
            "type": "object",
            "description": "`null`이면 저장된 locales 제거",
            "additionalProperties": true
          },
          "audioUrl": {
            "type": "string"
          },
          "sequence": {
            "type": "number"
          }
        }
      },
      "CreateMaterialDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "GRAMMAR_TABLE"
          },
          "sequence": {
            "type": "number"
          },
          "isExtra": {
            "type": "boolean"
          },
          "contentText": {
            "type": "object",
            "description": "JSON 본문 (예: 문법 표, 리스닝 audioUrl 등)"
          }
        },
        "required": [
          "type",
          "sequence",
          "contentText"
        ]
      },
      "PatchMaterialDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "sequence": {
            "type": "number"
          },
          "isExtra": {
            "type": "boolean"
          },
          "contentText": {
            "type": "object"
          }
        }
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "MCQ"
          },
          "questionText": {
            "type": "string"
          },
          "options": {
            "type": "object",
            "description": "선택지 등 JSON"
          },
          "answer": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "questionText",
          "options",
          "answer"
        ]
      },
      "PatchQuestionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "options": {
            "type": "object"
          },
          "answer": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          }
        }
      },
      "AdminAudioPresignedDto": {
        "type": "object",
        "properties": {
          "fileExtension": {
            "type": "string",
            "example": "mp3"
          },
          "contentType": {
            "type": "string",
            "example": "audio/mpeg"
          }
        },
        "required": [
          "fileExtension",
          "contentType"
        ]
      },
...
    }
  }
}